backend/requirements.txt
fastapi
uvicorn
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
chromadb
python-multipart
pypdf
sentence-transformers
langchain-groq==0.2.0
python-dotenv
groq 